requirements.txt
flask
crypt
pymongo
python-dotenv